Projekt aufgeräumt, Klassen in die richtigen Packages refactored
