+++ app/src/androidTest/java/at/aau/serg/sdlapp/ui/PlayerStatsOverlayTest.kt
@@ -3,6 +3,8 @@ package at.aau.serg.sdlapp.ui
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
+import at.aau.serg.sdlapp.model.player.PlayerModell
+import at.aau.serg.sdlapp.ui.PlayerOverlay.PlayerStatsOverlay
 import org.junit.Rule
 import org.junit.Test
