+++ app/src/main/java/at/aau/serg/sdlapp/ui/PlayerOverlay/AllPlayerStatsScreen.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.sdlapp.ui
+package at.aau.serg.sdlapp.ui.PlayerOverlay
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -8,6 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import at.aau.serg.sdlapp.network.viewModels.PlayerViewModel
 
 @Composable
 fun AllPlayerStatsScreen(viewModel: PlayerViewModel) {
