+++ app/src/main/java/at/aau/serg/sdlapp/ui/PlayerOverlay/PlayerStatsOverlay.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.sdlapp.ui
+package at.aau.serg.sdlapp.ui.PlayerOverlay
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -17,6 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import at.aau.serg.sdlapp.model.player.PlayerModell
 
 @Composable
 fun PlayerStatsOverlay(player: PlayerModell) {
