+++ app/src/main/java/at/aau/serg/sdlapp/ui/PlayerOverlay/PlayerStatsOverlayScreen.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.sdlapp.ui
+package at.aau.serg.sdlapp.ui.PlayerOverlay
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -15,6 +15,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import at.aau.serg.sdlapp.network.viewModels.PlayerViewModel
 
 @Composable
 fun PlayerStatsOverlayScreen(
